template override

+++ page/overridetemplate.php
@@ -20,14 +20,21 @@ class page_overridetemplate extends \Page {
 		parent::init();
 
 		if(!$this->app->auth->isLoggedIn()) return;
-		
+		if(!$_GET['xepan-tool-to-clone']){
+			return $this->add('View')->set('Please select a tool');
+		}	
 
 		$tool = $this->add($_GET['xepan-tool-to-clone']);
+		
+		if(!$tool->teplateOverridable){
+			$this->add('View')->set('You cannot override template for this tool');
+			return;
+		}
 		$original_path = $tool->template->origin_filename;
 		$tool->destroy();
 
 		$specific_path = substr($original_path, strpos($original_path,'/templates/')+strlen('/templates/'));
-		$override_path = $this->app->pathfinder->base_location->base_path.'/'.$this->app->current_website_name.'/www/'.$specific_path;
+		$override_path = $this->app->pathfinder->base_location->base_path.'/websites/'.$this->app->current_website_name.'/www/'.$specific_path;
 
 		if(file_exists($override_path)){
 			$this->add('View')->set('File allrealy overrided at "'. $specific_path.'", Please remove this file and click again to reset');
